Temporarily commented out all BQPD tests
