+++ uno/ingredients/inequality_handling_methods/InequalityHandlingMethodFactory.cpp
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project directory for details.
 
 #include <string>
-#include "Subproblem.hpp"
+#include "InequalityHandlingMethod.hpp"
 #include "SubproblemFactory.hpp"
 #include "ingredients/subproblems/inequality_constrained_methods/QPSubproblem.hpp"
 #include "ingredients/subproblems/inequality_constrained_methods/LPSubproblem.hpp"
@@ -12,7 +12,7 @@
 #include "options/Options.hpp"
 
 namespace uno {
-   std::unique_ptr<Subproblem> SubproblemFactory::create(size_t number_variables, size_t number_constraints, size_t number_objective_gradient_nonzeros,
+   std::unique_ptr<InequalityHandlingMethod> SubproblemFactory::create(size_t number_variables, size_t number_constraints, size_t number_objective_gradient_nonzeros,
          size_t number_jacobian_nonzeros, size_t number_hessian_nonzeros, const Options& options) {
       const std::string subproblem_strategy = options.get_string("subproblem");
       // active-set methods
