+++ uno/ingredients/inequality_handling_methods/InequalityHandlingMethodFactory.hpp
@@ -10,11 +10,11 @@
 namespace uno {
    // forward declaration
    class Options;
-   class Subproblem;
+   class InequalityHandlingMethod;
 
    class SubproblemFactory {
       public:
-         static std::unique_ptr<Subproblem> create(size_t number_variables, size_t number_constraints, size_t number_objective_gradient_nonzeros,
+         static std::unique_ptr<InequalityHandlingMethod> create(size_t number_variables, size_t number_constraints, size_t number_objective_gradient_nonzeros,
                size_t number_jacobian_nonzeros, size_t number_hessian_nonzeros, const Options& options);
 
          static std::vector<std::string> available_strategies();
