+++ uno/ingredients/inequality_handling_methods/inequality_constrained_methods/LPSubproblem.hpp
@@ -25,9 +25,6 @@ namespace uno {
    private:
       // pointer to allow polymorphism
       const std::unique_ptr<LPSolver> solver; /*!< Solver that solves the subproblem */
-      const SymmetricMatrix<size_t, double> zero_hessian;
-
-      void evaluate_functions(const OptimizationProblem& problem, Iterate& current_iterate, const WarmstartInformation& warmstart_information);
    };
 } // namespace
