+++ uno/ingredients/inequality_handling_methods/interior_point_methods/PrimalDualInteriorPointSubproblem.hpp
@@ -48,7 +48,7 @@ namespace uno {
 
    protected:
       SparseVector<double> objective_gradient; /*!< Sparse Jacobian of the objective */
-      std::vector<double> constraints; /*!< Constraint values (size \f$m)\f$ */
+      Vector<double> constraints; /*!< Constraint values (size \f$m)\f$ */
       RectangularMatrix<double> constraint_jacobian; /*!< Sparse Jacobian of the constraints */
 
       SymmetricIndefiniteLinearSystem<double> augmented_system;
